Backend/requirements.txt
NumPy
Pandas
Flask
Flask-CORS
Flask-SQLAlchemy
Pillow
Flask-Login 
Flask-Bcrypt
torch
torchvision
numpy